Naming consistency: `Change` (imperative) vs `Changed` (past/query-form)
